+++ bin/trin/src/main.rs
@@ -2,7 +2,7 @@
 
 use ethportal_api::types::network_spec::set_network_spec;
 use tracing::error;
-use trin::{cli::TrinConfig, run_trin_from_trin_config};
+use trin::{builder::run_trin_from_trin_config, cli::TrinConfig};
 use trin_utils::log::init_tracing_logger;
 
 #[tokio::main]
@@ -16,11 +16,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         .await
         .expect("failed to pause until ctrl-c");
 
-    let Some(rpc_handle) = trin_handle.rpc_server_handle else {
-        return Err("RPC server handle wasn't initialized".into());
-    };
-
-    if let Err(err) = rpc_handle.stop() {
+    if let Err(err) = trin_handle.rpc_server_handle.stop() {
         error!(err = %err, "Failed to close RPC server")
     }
